Further Implemented Backend Logic
Backend Logic for handling account creation and account logins are now implemented. Account logins are checked to ensure that they are valid, IDs for the new users are created and ensured that they are none repeating.

I need to still implement account creation logic regarding repeated emails (trying to make a new account with an email that is already in the system).